Add "New Language" template for TypeScript and moved old one as JS version.
With this change, ext-language becomes `ext-language-js` which now has
a sibling called `ext-language-ts`.

TypeScript-based template would be useful and less confusing when creating
a new language support with appropriate .vscode settings. When we bring
TypeScript sources into a project which was based on the JavaScript
template, It is very easy to get stuck at problems like this:
http://stackoverflow.com/questions/35606423/visual-studio-code-debug-node-js-through-typescript/