add notebook output renderer generator
This is using proposed APIs, so I hid it behind an environment variable
for now--you need to use `INCLUDE_PROPOSED=1 yo code` to see it.

There are quite a few prompts for it. We could possibly skip the
renderer ID and display name prompts and just use the name and
displayname of the extension instead if this is a concern.